[refactor] Move presence broadcasts and deferred admission out
Both sections read the shared registries and little else once those were
extracted: presence needs three collaborators, deferred admission five, where
before the registries moved they needed eleven and ten. Each takes them
through init() and keeps its own state — the presence heartbeat timer, the
two in-flight guards — so destroySwarm calls their resets instead of naming
their internals. swarm.js re-exports the five names worker/main.js and
boot.js import, so no caller changes.
